+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [1.2.32] - Unreleased
 - Free singleshot heap-allocated trait-object(callbacks) in app::awake_callback() and add_timeout() by @mosolovsa.
 - Fix TiledImage refcount.
+- Add experimental GlWidgetWindow.
 
 ## [1.2.31] - 2022-02-19
 - Add dialog::message_icon().